move helper methods to trait service
