+++ src/main/java/org/breedinginsight/services/TraitService.java
@@ -22,10 +22,19 @@ import io.micronaut.http.exceptions.HttpStatusException;
 import io.micronaut.http.server.exceptions.HttpServerException;
 import io.micronaut.http.server.exceptions.InternalServerException;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.tuple.Pair;
 import org.apache.commons.text.WordUtils;
-import org.brapi.v2.model.pheno.BrAPIObservation;
+import org.brapi.client.v2.model.exceptions.ApiException;
+import org.brapi.v2.model.BrAPIOntologyReference;
+import org.brapi.v2.model.core.BrAPIStudy;
+import org.brapi.v2.model.core.BrAPITrial;
+import org.brapi.v2.model.pheno.*;
 import org.breedinginsight.api.auth.AuthenticatedUser;
 import org.breedinginsight.api.model.v1.response.ValidationErrors;
+import org.breedinginsight.brapi.v2.constants.BrAPIAdditionalInfoFields;
+import org.breedinginsight.brapi.v2.services.BrAPITrialService;
+import org.breedinginsight.brapps.importer.services.ExternalReferenceSource;
 import org.breedinginsight.dao.db.enums.DataType;
 import org.breedinginsight.dao.db.tables.pojos.MethodEntity;
 import org.breedinginsight.dao.db.tables.pojos.ProgramSharedOntologyEntity;
@@ -37,6 +46,8 @@ import org.breedinginsight.services.exceptions.DoesNotExistException;
 import org.breedinginsight.services.exceptions.ValidatorException;
 import org.breedinginsight.services.validators.TraitValidatorError;
 import org.breedinginsight.services.validators.TraitValidatorService;
+import org.breedinginsight.utilities.Utilities;
+import org.jetbrains.annotations.NotNull;
 import org.jooq.DSLContext;
 
 import javax.inject.Inject;
@@ -60,14 +71,15 @@ public class TraitService {
     private TraitValidatorService traitValidator;
     private DSLContext dsl;
     private TraitValidatorError traitValidatorError;
-
+    private BrAPITrialService trialService;
+    private String referenceSource;
     private final static String FAVORITES_TAG = "favorites";
 
     @Inject
     public TraitService(TraitDAO traitDao, MethodDAO methodDao, ScaleDAO scaleDao, ObservationDAO observationDao, ProgramService programService,
                         ProgramOntologyService programOntologyService, ProgramObservationLevelService programObservationLevelService,
                         UserService userService, TraitValidatorService traitValidator, DSLContext dsl, TraitValidatorError traitValidatorError,
-                        ProgramOntologyDAO programOntologyDAO) {
+                        ProgramOntologyDAO programOntologyDAO, BrAPITrialService trialService, String referenceSource) {
         this.traitDAO = traitDao;
         this.methodDAO = methodDao;
         this.scaleDAO = scaleDao;
@@ -80,6 +92,8 @@ public class TraitService {
         this.dsl = dsl;
         this.traitValidatorError = traitValidatorError;
         this.programOntologyDAO = programOntologyDAO;
+        this.trialService = trialService;
+        this.referenceSource = referenceSource;
     }
 
     public List<Trait> getByProgramId(UUID programId, boolean getFullTrait) throws DoesNotExistException {
@@ -482,5 +496,183 @@ public class TraitService {
         }
 
         return traitDAO.getTraitsByTraitName(programId, names.stream().map(name -> Trait.builder().observationVariableName(name).build()).collect(Collectors.toList()));
+    }
+
+    public List<Trait> getBrAPIObservationVariablesForExperiment(
+            UUID programId,
+            Optional<String> experimentId,
+            Optional<String> environmentId
+    ) throws DoesNotExistException, ApiException {
+        log.debug(String.format("fetching variables for experiment.  expId: %s, envId: %s ",
+                experimentId.orElse(""), environmentId.orElse("")));
+        Optional<Program> program = programService.getById(programId);
+        if(program.isEmpty()) {
+            throw new DoesNotExistException("Could not find program: " + programId);
+        }
+        UUID expId;
+        if(experimentId.isPresent()) {
+            expId = UUID.fromString(experimentId.get());
+        } else {
+            UUID envId = UUID.fromString(environmentId.get());
+            BrAPIStudy environment = trialService.getEnvironment(program.get(), envId);
+            expId = UUID.fromString(Utilities.getExternalReference(environment.getExternalReferences(),
+                    Utilities.generateReferenceSource(referenceSource, ExternalReferenceSource.TRIALS)).get().getReferenceId());
+        }
+
+        BrAPITrial experiment = trialService.getExperiment(program.get(), expId);
+        if(experiment
+                .getAdditionalInfo().getAsJsonObject()
+                .has(BrAPIAdditionalInfoFields.OBSERVATION_DATASET_ID)) {
+            String obsDatasetId = experiment
+                    .getAdditionalInfo().getAsJsonObject()
+                    .get(BrAPIAdditionalInfoFields.OBSERVATION_DATASET_ID).getAsString();
+            return trialService.getDatasetObsVars(obsDatasetId, program.get());
+        }
+
+        return new ArrayList<>();
+    }
+
+    @NotNull
+    public List<BrAPIObservationVariable> filterVariables(List<Trait> programTraits,
+                                                           Optional<String> observationVariableDbId,
+                                                           Optional<String> observationVariableName,
+                                                           Optional<String> traitClass,
+                                                           Optional<String> methodDbId,
+                                                           Optional<String> methodName,
+                                                           Optional<String> scaleDbId,
+                                                           Optional<String> scaleName,
+                                                           Optional<String> traitDbId,
+                                                           Optional<String> traitName,
+                                                           Optional<String> ontologyDbId) throws DoesNotExistException {
+        log.debug("filtering variables:\n" +
+                "observationVariableDbId: " + observationVariableDbId + "\n" +
+                "observationVariableName: " + observationVariableName + "\n" +
+                "traitClass: " + traitClass + "\n" +
+                "methodDbId: " + methodDbId + "\n" +
+                "methodName: " + methodName + "\n" +
+                "scaleDbId: " + scaleDbId + "\n" +
+                "scaleName: " + scaleName + "\n" +
+                "traitDbId: " + traitDbId + "\n" +
+                "traitName: " + traitName + "\n" +
+                "ontologyDbId: " + ontologyDbId);
+
+        return programTraits.stream()
+                .filter(trait -> {
+                    boolean matches = true;
+
+                    Map<String, Pair<Optional<String>, String>> filterParams = new HashMap<>();
+                    filterParams.put("observationVariableDbId",
+                            Pair.of(observationVariableDbId,
+                                    trait.getId()
+                                            .toString()));
+                    filterParams.put("observationVariableName", Pair.of(observationVariableName, trait.getObservationVariableName()));
+                    filterParams.put("traitClass", Pair.of(traitClass, trait.getTraitClass()));
+                    filterParams.put("methodDbId",
+                            Pair.of(methodDbId,
+                                    trait.getMethodId()
+                                            .toString()));
+                    filterParams.put("methodName",
+                            Pair.of(methodName,
+                                    trait.getMethod()
+                                            .getDescription()));
+                    filterParams.put("scaleDbId",
+                            Pair.of(scaleDbId,
+                                    trait.getScale()
+                                            .getId()
+                                            .toString()));
+                    filterParams.put("scaleName",
+                            Pair.of(scaleName,
+                                    trait.getScale()
+                                            .getScaleName()));
+                    filterParams.put("traitDbId",
+                            Pair.of(traitDbId,
+                                    trait.getId()
+                                            .toString()));
+                    filterParams.put("traitName", Pair.of(traitName, trait.getObservationVariableName()));
+                    filterParams.put("ontologyDbId",
+                            Pair.of(ontologyDbId,
+                                    trait.getProgramOntologyId()
+                                            .toString()));
+
+                    for (Map.Entry<String, Pair<Optional<String>, String>> filter : filterParams.entrySet()) {
+                        if (filter.getValue().getLeft().isPresent()) {
+                            log.debug("filtering traits by: " + filter.getKey());
+                            matches = StringUtils.equals(filter.getValue()
+                                            .getLeft()
+                                            .get(),
+                                    filter.getValue()
+                                            .getRight());
+                        }
+                        if (!matches) {
+                            break;
+                        }
+                    }
+
+                    return matches;
+                })
+                .map(this::convertToBrAPI)
+                .collect(Collectors.toList());
+    }
+
+    public BrAPIObservationVariable convertToBrAPI(Trait trait) {
+        BrAPIOntologyReference brAPIOntologyReference = new BrAPIOntologyReference().ontologyDbId(trait.getProgramOntologyId()
+                .toString());
+        String status = trait.getActive() ? "active" : "inactive";
+        List<String> synonyms = prepSynonyms(trait);
+        return new BrAPIObservationVariable().observationVariableDbId(trait.getId().toString())
+                .observationVariableName(trait.getObservationVariableName())
+                .defaultValue(trait.getDefaultValue())
+                .status(status)
+                .synonyms(synonyms)
+                .trait(new BrAPITrait().ontologyReference(brAPIOntologyReference)
+                        .traitName(trait.getObservationVariableName())
+                        .traitDbId(trait.getId().toString())
+                        .entity(trait.getEntity())
+                        .attribute(trait.getAttribute())
+                        .status(status)
+                        .synonyms(synonyms))
+                .method(new BrAPIMethod().ontologyReference(brAPIOntologyReference)
+                        .methodDbId(trait.getMethod().getId().toString())
+                        .methodClass(trait.getMethod().getMethodClass())
+                        .description(trait.getMethod().getDescription())
+                        .formula(trait.getMethod().getFormula()))
+                .scale(new BrAPIScale().ontologyReference(brAPIOntologyReference)
+                        .scaleDbId(trait.getScale().getId().toString())
+                        .scaleName(trait.getScale().getScaleName())
+                        .dataType(BrAPITraitDataType.fromValue(trait.getScale().getDataType().getLiteral()))
+                        .decimalPlaces(trait.getScale().getDecimalPlaces())
+                        .validValues(new BrAPIScaleValidValues().max(trait.getScale().getValidValueMax())
+                                .min(trait.getScale().getValidValueMin())
+                                .categories(trait.getScale().getCategories())));
+    }
+
+    /**
+     * Create a list of synonyms, and ensure that there the first element matches the name of the trait<br><br>
+     * This is primarily needed to ensure any system using the first synonym as a display shows the actual name of the ontology term
+     * @param trait
+     * @return list of synonyms with at least one value (the name of the trait)
+     */
+    private List<String> prepSynonyms(Trait trait) {
+        List<String> preppedSynonyms = new ArrayList<>();
+        if(trait.getSynonyms() != null) {
+            preppedSynonyms = trait.getSynonyms();
+            int traitNameIdx = -1;
+            for(int i = 0; i < preppedSynonyms.size(); i++) {
+                if(preppedSynonyms.get(i).equals(trait.getObservationVariableName())) {
+                    traitNameIdx = i;
+                    break;
+                }
+            }
+            if(traitNameIdx > -1) {
+                String temp = preppedSynonyms.get(traitNameIdx);
+                preppedSynonyms.set(traitNameIdx, preppedSynonyms.get(0));
+                preppedSynonyms.set(0, temp);
+            } else {
+                preppedSynonyms.add(0, trait.getObservationVariableName());
+            }
+        } else {
+            preppedSynonyms.add(trait.getObservationVariableName());
+        }
+        return preppedSynonyms;
     }
 }
